+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,12 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import enhanced_preferences
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  EnhancedPreferencesPlugin.register(with: registry.registrar(forPlugin: "EnhancedPreferencesPlugin"))
+}